Use `this.import()` to import JS assets from the `inputmask` package
Using `ember-cli-node-assets` is incompatible with Embroider and can also cause issues when used in a nested addon. Ember CLI has provided an `import()` method for quite some time now, that can take care of a lot of the the same functionality, but is better supported in the edge cases mentioned above.

See also https://github.com/adopted-ember-addons/ember-pikaday/pull/234 for another example of this.